+++ Sugarboy2.0/Assets/Prefabs/GameManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d72fd094bf470b04ab03cb08fa03844c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
