+++ Sugarboy2.0/Assets/Prefabs/SpawnPoint.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1c364ba22b02a934b996e07d40bfc13b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
