+++ Sugarboy2.0/Assets/Scripts/PlayerMovement.cs
@@ -32,7 +32,7 @@ public class PlayerMovement : MonoBehaviour
 
     private void Update()
     {
-        if (Input.GetKeyDown(KeyCode.Q))
+        if (Input.GetKeyDown(KeyCode.Q) && GameManager.Instance.AllowedClones > 0)
         {
             this.tag = "Clone";
         }
@@ -66,4 +66,6 @@ public class PlayerMovement : MonoBehaviour
 
         return hitDetected;
     }
+
+    //TODO -> last clone moves with the player
 }
