+++ Sugarboy2.0/Assets/Scripts/SpawnClones.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f6ea3806383bf884abe328cea882db5a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
